Icon+label+HeroCollection+refactoring
