+++ Lab/PhotoCell.swift
@@ -1,0 +1,52 @@
+//
+//  PhotoCell.swift
+//  Lab
+//
+//  Created by Dmitrii Shelema on 11.03.2024.
+//
+import UIKit
+
+class PhotoCell: UICollectionViewCell {
+    let imageView = UIImageView()
+    let nameLabel = UILabel()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupUI()
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func setupUI() {
+        imageView.contentMode = .scaleAspectFill
+        contentView.addSubview(imageView)
+        
+        // Настройка ограничений для imageView
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            imageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            imageView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            imageView.topAnchor.constraint(equalTo: contentView.topAnchor),
+            imageView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor)
+        ])
+        
+        // UILabel для отображения имени персонажа
+        nameLabel.textColor = .white
+        nameLabel.textAlignment = .left
+        nameLabel.font = UIFont.systemFont(ofSize: 28, weight: .bold)
+        contentView.addSubview(nameLabel)
+        nameLabel.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            nameLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 30),
+            nameLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            nameLabel.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
+            nameLabel.heightAnchor.constraint(equalToConstant: 80)
+        ])
+        
+        // Закругление углов imageView
+        imageView.layer.cornerRadius = 10
+        imageView.layer.masksToBounds = true
+    }
+}
